rename info-server
