+++ pynumaflow/tests/function/test_info_server.py
@@ -3,12 +3,12 @@ import unittest
 from unittest import mock
 
 from pynumaflow.tests.function.testing_utils import read_info_server
-from pynumaflow.info.info_server import (
+from pynumaflow.info.server import (
     get_sdk_version,
     write as info_server_write,
     get_metadata_env,
 )
-from pynumaflow.info.info_types import (
+from pynumaflow.info.types import (
     ServerInfo,
     Protocol,
     Language,
